backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.5.14
DateTime==5.2
distlib==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.2
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
gensim==4.3.2
gpt4all==1.0.12
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.312
langchainhub==0.1.13
langcodes==3.3.0
langsmith==0.0.43
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.26.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.1.1
pathy==0.10.2
pdf2image==1.16.3
pdfkit==1.0.0
Pillow==10.0.0
pinecone-client==2.2.4
platformdirs==3.10.0
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
replicate==0.15.2
reportlab==4.0.4
repoze.lru==0.7
requests==2.31.0
Routes==2.5.1
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.19
srsly==2.4.8
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
types-requests==2.31.0.8
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
virtualenv==20.24.5
wasabi==1.1.2
weasel==0.3.2
win32-setctime==1.1.0
wkhtmltopdf==0.2
yarl==1.9.2
zope.interface==6.0